Add flex properties to body
Used flex to stick the header on the head and the footer on the foot.
Added in a fill block above the footer to fill in the gap between the
bottom of the screen and the bottom of the current content. Can probably
remove later after everything is filled in.